requirements.txt

# Progess bars
tqdm>=4.61.2

# Table visualisation
tabulate>=0.8.9

# UI Themes
dacite>=1.6.0  # Convert dictionaries (-> json / toml) to dataclasses
toml>=0.10.2

# Plant search feature in UI
fuzzyset2>=0.1.1

# Web crawler / scraper
requests>=2.24.0
bs4
beautifulsoup4>=4.8.2

# Text processing
# You still have to run python -m spacy download en_core_web_sm
spacy>=3.1.1
